requirements.txt
matplotlib==3.9.1.post1
numpy==2.0.1
pandas==2.2.2
PyYAML==6.0.2
regex==2024.7.24
scipy==1.14.0
seaborn==0.13.2
openpyxl==3.1.5
XlsxWriter==3.2.0
tqdm==4.66.2